RPC client-js: abstract from tut-6-rpc
