--- examples/tut-6-rpc/client-js/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@reach-sh/tut-6-rpc-client-js",
-  "type": "module",
-  "dependencies": {
-    "wait-port": "^0.2.9",
-    "bent": "^7.3.12"
-  },
-  "author": "reach.sh",
-  "license": "Apache-2.0",
-  "scripts": {
-    "lint": "eslint --ignore-path .gitignore --ext .mjs .",
-    "index": "node --experimental-modules --unhandled-rejections=strict index.mjs"
-  }
-}
