+++ js/JS_IMAGES
@@ -5,6 +5,8 @@ STDLIB_IMAGE="reachsh/stdlib"
 RUNNER_IMAGE="reachsh/runner"
 REACT_RUNNER_IMAGE="reachsh/react-runner"
 RPC_SERVER_IMAGE="reachsh/rpc-server"
+RPC_CLIENT_IMAGE="reachsh/rpc-client"
+RPC_CLIENT_RUNNER_IMAGE="reachsh/rpc-client-runner"
 
 REACH_VERSION=$(VERSION)
 NODE_VERSION="12.18.3"
