+++ js/js-deps/package.json
@@ -18,7 +18,9 @@
     "ethers": "^5.0.19",
     "hi-base32": "^0.5.0",
     "wait-port": "^0.2.9",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "wait-port": "^0.2.9",
+    "bent": "^7.3.12"
   },
   "optionalDependencies": {
     "ganache-core": "^2.11.2"
