+++ js/rpc-client-runner/package.mo.json
@@ -1,0 +1,22 @@
+{
+  "name": "@reach-sh/rpc-client-runner",
+  "description": "JSON-RPC client runner for your DApp + the Reach standard library",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/reach-sh/reach-lang.git",
+    "directory": "js"
+  },
+  "type": "module",
+  "dependencies": {
+    "@reach-sh/rpc-client": "{{REACH_VERSION}}"
+  },
+  "author": "reach.sh",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/reach-sh/reach-lang/issues"
+  },
+  "homepage": "https://github.com/reach-sh/reach-lang#readme",
+  "scripts": {
+    "app": "node --experimental-modules --unhandled-rejections=strict index.mjs"
+  }
+}
