+++ js/rpc-client/Dockerfile
@@ -1,0 +1,5 @@
+ARG  RUNNER_VERSION
+
+# hadolint ignore=DL3006
+FROM $RUNNER_VERSION
+COPY package.json index.mjs /app/node_modules/@reach-sh/rpc-client/
