+++ js/rpc-client/package.mo.json
@@ -1,0 +1,20 @@
+{
+  "name": "@reach-sh/rpc-client",
+  "description": "JSON-RPC client to your DApp + the Reach standard library",
+  "main": "./index.mjs",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/reach-sh/reach-lang.git",
+    "directory": "js"
+  },
+  "type": "module",
+  "dependencies": {
+    "@reach-sh/stdlib": "{{REACH_VERSION}}"
+  },
+  "author": "reach.sh",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/reach-sh/reach-lang/issues"
+  },
+  "homepage": "https://github.com/reach-sh/reach-lang#readme"
+}
